Move functions around based on feedback.
