+++ packages/story-editor/src/components/panels/shared/captionsPanelContent.js
@@ -44,7 +44,7 @@ import Tooltip from '../../tooltip';
 import { useConfig } from '../../../app/config';
 import { MULTIPLE_DISPLAY_VALUE } from '../../../constants';
 import { focusStyle } from './styles';
-import HotlinkModal from './hotLinkModel';
+import HotlinkModal from './hotlinkModal';
 
 const InputRow = styled.div`
   display: flex;
