--- packages/story-editor/src/utils/getErrorMessage.js
@@ -1,39 +1,0 @@
-/*
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * External dependencies
- */
-import { __, sprintf } from '@googleforcreators/i18n';
-
-function getErrorMessage(code, description) {
-  switch (code) {
-    case 'rest_invalid_param':
-    case 'rest_invalid_url':
-      return __('Invalid link.', 'web-stories');
-    case 'rest_invalid_ext':
-      return sprintf(
-        /* translators: %s is the description with allowed file extensions. */
-        __('Invalid link. %s', 'web-stories'),
-        description
-      );
-    default:
-      return __(
-        'Media failed to load. Please ensure the link is valid and the site allows linking from external sites.',
-        'web-stories'
-      );
-  }
-}
-export default getErrorMessage;
